server/requirements.txt
autopep8==1.5
Click==7.0
colorama==0.4.3
cursor==1.3.4
deepspeech==0.6.1
dnspython==1.16.0
entrypoints==0.3
eventlet==0.25.1
ffmpeg-python==0.2.0
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SocketIO==4.2.1
future==0.18.2
greenlet==0.4.15
gunicorn==20.0.4
halo==0.0.28
itsdangerous==1.1.0
Jinja2==2.11.1
log-symbols==0.0.14
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
numpy==1.18.1
PyAudio==0.2.11
pycodestyle==2.5.0
pyflakes==2.1.1
python-engineio==3.11.2
python-socketio==4.4.0
scipy==1.4.1
six==1.14.0
spinners==0.0.23
termcolor==1.1.0
Wave==0.0.2
webrtcvad==2.0.10
Werkzeug==1.0.0
